added process it control for processing images

+++ core/utils/image.py
@@ -22,26 +22,26 @@ def create_default_image(image, process_it=False) -> Image:
         is_processed=process_it
     )
     writed, final_filename = process_img(image_obj)
-    if writed:
-        # Open processed image for saving it to model
-        processed_img = PImage.open(f'media/{final_filename}')
-        bytes_img = BytesIO()
-        processed_img.save(bytes_img, 'JPEG')
+    if process_it:
+        if writed:
+            # Open processed image for saving it to model
+            processed_img = PImage.open(f'media/{final_filename}')
+            bytes_img = BytesIO()
+            processed_img.save(bytes_img, 'JPEG')
 
-        # craete model from opened image
-        processed_obj = ProcessedImage.objects.create(
-            processed_image=ImageFile(bytes_img, final_filename)
-        )
+            # craete model from opened image
+            processed_obj = ProcessedImage.objects.create(
+                processed_image=ImageFile(bytes_img, final_filename)
+            )
 
-        # Edit default image for saying it has processed image
-        image_obj.related_img = processed_obj
-        image_obj.is_processed = True
-        image_obj.save()
+            # Edit default image for saying it has processed image
+            image_obj.related_img = processed_obj
+            image_obj.is_processed = True
+            image_obj.save()
 
-        print(f'Successfully created processed img: {processed_obj}')
-        print('deleting temp image')
-        # delete unneccessary image cuz it is duplicated
-        os.remove(f'media/{final_filename}')
-        return image_obj, processed_obj
-
+            print(f'Successfully created processed img: {processed_obj}')
+            print('deleting temp image')
+            # delete unneccessary image cuz it is duplicated
+            os.remove(f'media/{final_filename}')
+            return image_obj, processed_obj
     return image_obj, None
